requirements.txt
ansible==4.8.0
ansible-core==2.11.6
cffi==1.15.0
cryptography==35.0.0
django-environ==0.8.1
gitdb==4.0.9
GitPython==3.1.24
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.2
passlib==1.7.4
pycparser==2.21
pyparsing==2.4.7
PyYAML==6.0
resolvelib==0.5.4
smmap==5.0.0
typing_extensions==4.0.1
